+++ plone/app/contenttypes/migration/migration.py
@@ -26,10 +26,18 @@ from plone.event.utils import default_timezone
 from plone.namedfile.file import NamedBlobFile
 from plone.namedfile.file import NamedBlobImage
 from z3c.relationfield import RelationValue
+from zope.component import adapter
+from zope.component import getAdapters
 from zope.component import getUtility
 from zope.event import notify
+from zope.interface import Interface
+from zope.interface import implementer
 from zope.intid.interfaces import IIntIds
 from zope.lifecycleevent import ObjectModifiedEvent
+
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 def migrate(portal, migrator):
@@ -131,7 +139,7 @@ def restoreReferences(portal):
         out += refs(obj)
         # backrefs
         out += backrefs(portal, obj)
-        #order
+        # order
         out += order(obj)
 
     return out
@@ -188,25 +196,75 @@ class ReferenceMigrator(object):
         self.new._relatedItemsOrder = self.old._relatedItemsOrder
 
 
-class ATCTBaseMigrator(CMFItemMigrator, ReferenceMigrator):
+class ICustomMigrator(Interface):
+    """Adapter implementer interface for custom migrators.
+    Please note that you have to register named adapters in order to be able to
+    register multiple adapters to the same adaptee.
+    """
+    def migrate(old, new):
+        """Start the custom migraton.
+        :param old: The old content object.
+        :param new: The new content object.
+        """
+
+
+@implementer(ICustomMigrator)
+@adapter(Interface)
+class BaseCustomMigator(object):
+    """Base custom migration class. Does nothing.
+
+    You can use this as base class for your custom migrator adapters.
+    You might register it to some specific orginal content interface.
+    """
+    def __init__(self, context):
+        self.context = context
+
+    def migrate(self, old, new):
+        return
+
+
+class ATCTContentMigrator(CMFItemMigrator, ReferenceMigrator):
+    """Base for contentish ATCT
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(ATCTContentMigrator, self).__init__(*args, **kwargs)
+        logger.info(
+            "Migrating object %s" %
+            '/'.join(self.old.getPhysicalPath())
+        )
 
     def migrate_atctmetadata(self):
         field = self.old.getField('excludeFromNav')
         self.new.exclude_from_nav = field.get(self.old)
 
+    def migrate_custom(self):
+        """Get all ICustomMigrator registered migrators and run the migration.
+        """
+        for _, migrator in getAdapters((self.old, ), ICustomMigrator):
+            migrator.migrate(self.old, self.new)
 
-class ATCTContentMigrator(ATCTBaseMigrator,
-                          CMFItemMigrator,
-                          ReferenceMigrator):
-    """Base for contentish ATCT
-    """
 
-
-class ATCTFolderMigrator(ATCTBaseMigrator,
-                         CMFFolderMigrator,
-                         ReferenceMigrator):
+class ATCTFolderMigrator(CMFFolderMigrator, ReferenceMigrator):
     """Base for folderish ATCT
     """
+
+    def __init__(self, *args, **kwargs):
+        super(ATCTFolderMigrator, self).__init__(*args, **kwargs)
+        logger.info(
+            "Migrating object %s" %
+            '/'.join(self.old.getPhysicalPath())
+        )
+
+    def migrate_atctmetadata(self):
+        field = self.old.getField('excludeFromNav')
+        self.new.exclude_from_nav = field.get(self.old)
+
+    def migrate_custom(self):
+        """Get all ICustomMigrator registered migrators and run the migration.
+        """
+        for _, migrator in getAdapters((self.old, ), ICustomMigrator):
+            migrator.migrate(self.old, self.new)
 
 
 class DocumentMigrator(ATCTContentMigrator):
@@ -251,20 +309,12 @@ def migrate_files(portal):
     return migrate(portal, FileMigrator)
 
 
-class BlobFileMigrator(ATCTContentMigrator):
+class BlobFileMigrator(FileMigrator):
 
     src_portal_type = 'File'
     src_meta_type = 'ATBlob'
     dst_portal_type = 'File'
     dst_meta_type = None  # not used
-
-    def migrate_schema_fields(self):
-        old_file = self.old.getField('file').get(self.old)
-        filename = safe_unicode(old_file.filename)
-        namedblobfile = NamedBlobFile(contentType=old_file.content_type,
-                                      data=old_file.data,
-                                      filename=filename)
-        self.new.file = namedblobfile
 
 
 def migrate_blobfiles(portal):
@@ -292,22 +342,12 @@ def migrate_images(portal):
     return migrate(portal, ImageMigrator)
 
 
-class BlobImageMigrator(ATCTContentMigrator):
+class BlobImageMigrator(ImageMigrator):
 
     src_portal_type = 'Image'
     src_meta_type = 'ATBlob'
     dst_portal_type = 'Image'
     dst_meta_type = None  # not used
-
-    def migrate_schema_fields(self):
-
-        old_image = self.old.getField('image').get(self.old)
-        if old_image == '':
-            return
-        filename = safe_unicode(old_image.filename)
-        namedblobimage = NamedBlobImage(data=old_image.data,
-                                        filename=filename)
-        self.new.image = namedblobimage
 
 
 def migrate_blobimages(portal):
@@ -360,7 +400,7 @@ def migrate_newsitems(portal):
     return migrate(portal, NewsItemMigrator)
 
 
-class BlobNewsItemMigrator(DocumentMigrator):
+class BlobNewsItemMigrator(NewsItemMigrator):
     """ Migrator for NewsItems with blobs based on the implementation in
         https://github.com/plone/plone.app.blob/pull/2
     """
@@ -369,24 +409,6 @@ class BlobNewsItemMigrator(DocumentMigrator):
     src_meta_type = 'ATBlobContent'
     dst_portal_type = 'News Item'
     dst_meta_type = None  # not used
-
-    def migrate_schema_fields(self):
-        # migrate the text
-        super(BlobNewsItemMigrator, self).migrate_schema_fields()
-
-        # migrate the rest of the Schema
-        old_image = self.old.getField('image').get(self.old)
-        if old_image == '':
-            return
-        filename = safe_unicode(old_image.filename)
-        old_image_data = old_image.data
-        if safe_hasattr(old_image_data, 'data'):
-            old_image_data = old_image_data.data
-        namedblobimage = NamedBlobImage(data=old_image_data,
-                                        filename=filename)
-        self.new.image = namedblobimage
-        self.new.image_caption = safe_unicode(
-            self.old.getField('imageCaption').get(self.old))
 
 
 def migrate_blobnewsitems(portal):
@@ -399,6 +421,10 @@ class FolderMigrator(ATCTFolderMigrator):
     src_meta_type = 'ATFolder'
     dst_portal_type = 'Folder'
     dst_meta_type = None  # not used
+
+    def beforeChange_migrate_layout(self):
+        if self.old.getLayout() == 'atct_album_view':
+            self.old.setLayout('folder_album_view')
 
 
 def migrate_folders(portal):