+++ plone/app/contenttypes/tests/test_folder.py
@@ -81,13 +81,24 @@ class FolderFunctionalTest(unittest.TestCase):
         self.browser.getLink(url='http://nohost/plone/++add++Folder').click()
         self.browser.getControl(name='form.widgets.IDublinCore.title')\
             .value = "My folder"
+        self.browser.getControl(name='form.widgets.IShortName.id')\
+            .value = ""
         self.browser.getControl(name='form.widgets.IDublinCore.description')\
             .value = "This is my folder."
         self.browser.getControl('Save').click()
-
         self.assertTrue(self.browser.url.endswith('my-folder/view'))
         self.assertTrue('My folder' in self.browser.contents)
         self.assertTrue('This is my folder' in self.browser.contents)
+
+    def test_add_folder_with_shortname(self):
+        self.browser.open(self.portal_url)
+        self.browser.getLink(url='http://nohost/plone/++add++Folder').click()
+        self.browser.getControl(name='form.widgets.IDublinCore.title')\
+            .value = "My folder"
+        self.browser.getControl(name='form.widgets.IShortName.id')\
+            .value = "my-special-folder"
+        self.browser.getControl('Save').click()
+        self.assertTrue(self.browser.url.endswith('my-special-folder/view'))
 
 
 class FolderViewFunctionalTest(unittest.TestCase):
@@ -133,7 +144,7 @@ class FolderViewFunctionalTest(unittest.TestCase):
         self.assertTrue('My Folder' in self.browser.contents)
         self.assertTrue('Document 1' in self.browser.contents)
 
-    def test_atct_album_view(self):
-        self.browser.open(self.folder_url + '/atct_album_view')
+    def test_folder_album_view(self):
+        self.browser.open(self.folder_url + '/folder_album_view')
         self.assertTrue('My Folder' in self.browser.contents)
         self.assertTrue('Document 1' in self.browser.contents)