+++ plone/app/contenttypes/upgrades.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from Products.CMFCore.utils import getToolByName
+from plone.app.contenttypes.utils import DEFAULT_TYPES
 from plone.dexterity.interfaces import IDexterityFTI
 from zope.component import queryUtility
 
@@ -45,8 +47,8 @@ def update_fti(context):
 
 
 def enable_collection_behavior(context):
-    """Enable collection behavior on Collection.
-    """
+    """Enable collection behavior on Collection."""
+
     fti = queryUtility(
         IDexterityFTI,
         name='Collection'
@@ -61,9 +63,47 @@ def enable_collection_behavior(context):
 
 
 def migrate_to_richtext(context):
-    """update fti's to add RichText behaviors and remove old text-fields"""
+    """Update fti's to add RichText behaviors and remove old text-fields."""
 
+    # TODO: Don't reload the profile. Only change the settings.
     context.runImportStepFromProfile(
         'profile-plone.app.contenttypes:default',
         'typeinfo',
     )
+
+
+def migrate_album_view(context):
+    """Migrate atct_album_view to folder_album_view."""
+
+    # TODO: Don't reload the profile. Only change the settings.
+    context.runImportStepFromProfile(
+        'profile-plone.app.contenttypes:default',
+        'typeinfo',
+    )
+    catalog = getToolByName(context, 'portal_catalog')
+    search = catalog.unrestrictedSearchResults
+    for brain in search(portal_type='Folder'):
+        obj = brain.getObject()
+        current = context.getLayout()
+        if current == 'atct_album_view':
+            obj.setLayout('folder_album_view')
+
+
+def enable_shortname_behavior(context):
+    """Add IShortName to all types."""
+
+    behavior = 'plone.app.dexterity.behaviors.id.IShortName'
+    for type_id in DEFAULT_TYPES:
+        fti = queryUtility(
+            IDexterityFTI,
+            name=type_id
+        )
+        if fti is None:
+            continue
+
+        if behavior in fti.behaviors:
+            continue
+        behaviors = list(fti.behaviors)
+        behaviors.append(behavior)
+        behaviors = tuple(behaviors)
+        fti._updateProperty('behaviors', behaviors)